requirements.txt
websocket-client
requests